Iterate

+++ antigen_test/src/components/pancurses_mouse_component.rs
@@ -6,7 +6,7 @@ use antigen::{
 #[derive(Debug, Copy, Clone)]
 pub struct PancursesMouseComponent {
     position: IVector2,
-    button_mask_buffer: [i64; 2],
+    button_mask_buffer: [usize; 2],
 }
 
 impl<'a> PancursesMouseComponent {
@@ -21,7 +21,7 @@ impl<'a> PancursesMouseComponent {
         self.position
     }
 
-    pub fn get_button_mask(&self) -> i64 {
+    pub fn get_button_mask(&self) -> usize {
         self.button_mask_buffer[1]
     }
 
@@ -35,13 +35,13 @@ impl<'a> PancursesMouseComponent {
         self
     }
 
-    pub fn set_button_mask(&mut self, button_mask: i64) -> &mut Self {
+    pub fn set_button_mask(&mut self, button_mask: usize) -> &mut Self {
         self.button_mask_buffer[0] = self.button_mask_buffer[1];
         self.button_mask_buffer[1] = button_mask;
         self
     }
 
-    pub fn was_button_just_pressed(&self, button_mask: i64) -> bool {
+    pub fn was_button_just_pressed(&self, button_mask: usize) -> bool {
         self.button_mask_buffer[1] & button_mask > 0
             && self.button_mask_buffer[0] & button_mask == 0
     }
